[oersi] Depend on metafacture-core oersi branch via mavenLocal
Using source dependency here causes issues in the downstream
oersi-etl, which depends on metafacture-fix. There is probably a
way to solve that, but the whole oersi branch setup is temporary,
and for development it's actually useful to install and depend
via the local maven repository (to test changes in core or fix).